Sun now renders as part of the skybox Fixed some issues with the fullscreen effects not behaving properly following the implementation of MSAA Reenabled camera light

+++ DirectX 11 Engine/Core/Handlers/System Handlers/GraphicsHandler.h
@@ -53,7 +53,7 @@ private:
 	int m_CurrentDay;
 	int m_CurrentSeason;
 
-	std::vector<int> m_SeasonsSunHeightOffset;
+	std::vector<float> m_SeasonsSunHeightOffset;
 	std::map<int, std::string> m_SeasonNames;
 
 	std::vector<PointLight> m_PointLights;
